+++ src/mods/commands/verack.h
@@ -1,11 +1,13 @@
 #ifndef VERACK_H
 #define VERACK_H 1
 
+#include <stddef.h>
+
 #define VERACK_COMMAND "verack"
 
 typedef struct Verack *Verack;
 
 int verack_new(Verack);
-int verack_sizeof(void);
+size_t verack_sizeof(void);
 
 #endif
